Phase UI Update + Dashboard: Tailwind, dua shell, dashboard data interaktif
Dashboard data baru: halaman Data dengan ringkasan 8 stat, histogram harga dan tahun, merek/lokasi/transmisi, semua dengan tooltip hover/sentuh/fokus dan scroll-reveal (MOTION 2). Backend tambah GET /api/v1/stats/dataset (filter brand/location/machine/year, histogram 12 bucket) via listing_repository + stats_service + schemas/stats; listings auto-seed 21.553 baris saat startup bila kosong (SEED_ON_STARTUP/SEED_CSV_PATH, CSV di-mount read-only); Listing.id BigInteger ke Integer (auto-increment SQLite). pytest 9/9 termasuk stats kosong/filter.

Migrasi styling ke Tailwind v4 (@tailwindcss/vite, @theme dari palet calm, 3 kelas bantu di CSS). Satu aplikasi dua shell via hook useWindowSize (batas 768px): DesktopView sidebar statis + konten center-ke-layar untuk kolom sempit; MobileView header ringkas + hamburger drawer (Escape/overlay) + Bottom Navigation Bar + safe-area. Nav terpusat di navItems.jsx (ikon relevan berdokumen), PageContent bersama. Perbaikan mobile: header anti-wrap, label histogram invisible-anti-tabrakan, bar transmisi vertikal, tipografi/padding fluid (clamp + breakpoint sm). Footer diselaraskan ke kolom lede 34rem di kedua shell.

Uji: vite build lolos tiap iterasi, smoke stats/filter/auto-seed, preview 200, bundle tanpa hardcode/emoji/em dash, kontras AA, Delivery Gate Antislop PASS (mode During, ENERGY 1/RHYTHM 1/MOTION 2). Diketahui: ganti shell me-remount halaman (state form tidak terbawa), styles.css sempat hilang misterius dan ditulis ulang terverifikasi.

+++ backend/app/db/models.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import JSON, BigInteger, Boolean, DateTime, ForeignKey, Numeric, String, func
+from sqlalchemy import JSON, BigInteger, Boolean, DateTime, ForeignKey, Integer, Numeric, String, func
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 from sqlalchemy import Uuid
@@ -24,7 +24,7 @@ class Dataset(Base):
 
 class Listing(Base):
     __tablename__ = "listings"
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     dataset_id: Mapped[uuid.UUID] = mapped_column(Uuid, ForeignKey("datasets.id"))
     listing_type: Mapped[str] = mapped_column(String(16))
     brand: Mapped[str] = mapped_column(String(64))
@@ -65,7 +65,7 @@ class Prediction(Base):
     __tablename__ = "predictions"
     id: Mapped[uuid.UUID] = mapped_column(Uuid, primary_key=True, default=uuid.uuid4)
     model_version_id: Mapped[uuid.UUID] = mapped_column(Uuid, ForeignKey("model_versions.id"))
-    listing_id: Mapped[int | None] = mapped_column(BigInteger, ForeignKey("listings.id"), nullable=True)
+    listing_id: Mapped[int | None] = mapped_column(Integer, ForeignKey("listings.id"), nullable=True)
     predicted_price: Mapped[int] = mapped_column(BigInteger)
     input_payload: Mapped[dict] = mapped_column(PayloadJSON)
     shap_payload: Mapped[dict] = mapped_column(PayloadJSON)
